refactor: move shapes to a dedicated directory
Previously, elements related to shapes were located at the same place as geometry or cell elements. This was unclear
and didn't explicitly split the concepts.

This is only an internal change which has no impact for applications integration maxGraph.
The elements are still exported to the outside world in the same way as before.
