+++ packages/core/__tests__/view/shape/node/StencilShapeRegistry.test.ts
