+++ packages/core/src/view/GraphView.ts
@@ -21,7 +21,7 @@ import Rectangle from './geometry/Rectangle';
 import Dictionary from '../util/Dictionary';
 import EventSource from './event/EventSource';
 import EventObject from './event/EventObject';
-import RectangleShape from './geometry/node/RectangleShape';
+import RectangleShape from './shape/node/RectangleShape';
 import { NS_SVG } from '../util/Constants';
 import Client from '../Client';
 import InternalEvent from './event/InternalEvent';
@@ -30,12 +30,12 @@ import { getRotatedPoint, ptSegDistSq, relativeCcw, toRadians } from '../util/ma
 import { GlobalConfig } from '../util/config';
 import CellState from './cell/CellState';
 import UndoableEdit from './undoable_changes/UndoableEdit';
-import ImageShape from './geometry/node/ImageShape';
+import ImageShape from './shape/node/ImageShape';
 import InternalMouseEvent from './event/InternalMouseEvent';
 import Cell from './cell/Cell';
 import Image from './image/ImageBox';
 import CurrentRootChange from './undoable_changes/CurrentRootChange';
-import Shape from './geometry/Shape';
+import Shape from './shape/Shape';
 import Geometry from './geometry/Geometry';
 import ConnectionConstraint from './other/ConnectionConstraint';
 import type PopupMenuHandler from './plugins/PopupMenuHandler';
