+++ packages/core/src/view/handler/VertexHandler.ts
@@ -19,9 +19,9 @@ limitations under the License.
 import Rectangle from '../geometry/Rectangle';
 import { NONE } from '../../util/Constants';
 import InternalEvent from '../event/InternalEvent';
-import RectangleShape from '../geometry/node/RectangleShape';
-import ImageShape from '../geometry/node/ImageShape';
-import EllipseShape from '../geometry/node/EllipseShape';
+import RectangleShape from '../shape/node/RectangleShape';
+import ImageShape from '../shape/node/ImageShape';
+import EllipseShape from '../shape/node/EllipseShape';
 import Point from '../geometry/Point';
 import { getRotatedPoint, intersects, mod, toRadians } from '../../util/mathUtils';
 import Client from '../../Client';
@@ -31,7 +31,7 @@ import CellState from '../cell/CellState';
 import Image from '../image/ImageBox';
 import type Cell from '../cell/Cell';
 import type { CellHandle, Listenable, MouseListenerSet } from '../../types';
-import Shape from '../geometry/Shape';
+import Shape from '../shape/Shape';
 import InternalMouseEvent from '../event/InternalMouseEvent';
 import EdgeHandler from './EdgeHandler';
 import EventSource from '../event/EventSource';
