+++ packages/core/src/view/shape/ShapeRegistry.ts
