+++ packages/core/src/view/shape/edge/LineShape.ts
@@ -18,7 +18,7 @@ limitations under the License.
 
 import Shape from '../Shape';
 import AbstractCanvas2D from '../../canvas/AbstractCanvas2D';
-import Rectangle from '../Rectangle';
+import Rectangle from '../../geometry/Rectangle';
 import { ColorValue } from '../../../types';
 
 /**
