+++ packages/core/src/view/shape/node/DoubleEllipseShape.ts
@@ -16,9 +16,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import Rectangle from '../Rectangle';
+import Rectangle from '../../geometry/Rectangle';
 import Shape from '../Shape';
-import AbstractCanvas2D from '../../../view/canvas/AbstractCanvas2D';
+import AbstractCanvas2D from '../../canvas/AbstractCanvas2D';
 
 /**
  * Extends {@link Shape} to implement a double ellipse shape.
