+++ packages/core/src/view/shape/node/HexagonShape.ts
@@ -17,38 +17,43 @@ limitations under the License.
 */
 
 import ActorShape from './ActorShape';
+import Point from '../../geometry/Point';
+import { LINE_ARCSIZE } from '../../../util/Constants';
 import AbstractCanvas2D from '../../canvas/AbstractCanvas2D';
-import Rectangle from '../Rectangle';
 
 /**
- * Extends {@link ActorShape} to implement a cloud shape.
+ * Implementation of the hexagon shape.
  *
- * This shape is registered under `cloud` in {@link CellRenderer} when using {@link Graph} or calling {@link registerDefaultShapes}.
+ * This shape is registered under `hexagon` in {@link CellRenderer} when using {@link Graph} or calling {@link registerDefaultShapes}.
  *
  * @category Vertex Shapes
  */
-class CloudShape extends ActorShape {
-  constructor(bounds: Rectangle, fill: string, stroke: string, strokeWidth = 1) {
+class HexagonShape extends ActorShape {
+  constructor() {
     super();
-    this.bounds = bounds;
-    this.fill = fill;
-    this.stroke = stroke;
-    this.strokeWidth = strokeWidth;
   }
 
   /**
    * Draws the path for this shape.
    */
   redrawPath(c: AbstractCanvas2D, x: number, y: number, w: number, h: number) {
-    c.moveTo(0.25 * w, 0.25 * h);
-    c.curveTo(0.05 * w, 0.25 * h, 0, 0.5 * h, 0.16 * w, 0.55 * h);
-    c.curveTo(0, 0.66 * h, 0.18 * w, 0.9 * h, 0.31 * w, 0.8 * h);
-    c.curveTo(0.4 * w, h, 0.7 * w, h, 0.8 * w, 0.8 * h);
-    c.curveTo(w, 0.8 * h, w, 0.6 * h, 0.875 * w, 0.5 * h);
-    c.curveTo(w, 0.3 * h, 0.8 * w, 0.1 * h, 0.625 * w, 0.2 * h);
-    c.curveTo(0.5 * w, 0.05 * h, 0.3 * w, 0.05 * h, 0.25 * w, 0.25 * h);
-    c.close();
+    const arcSize = (this.style?.arcSize ?? LINE_ARCSIZE) / 2;
+
+    this.addPoints(
+      c,
+      [
+        new Point(0.25 * w, 0),
+        new Point(0.75 * w, 0),
+        new Point(w, 0.5 * h),
+        new Point(0.75 * w, h),
+        new Point(0.25 * w, h),
+        new Point(0, 0.5 * h),
+      ],
+      this.isRounded,
+      arcSize,
+      true
+    );
   }
 }
 
-export default CloudShape;
+export default HexagonShape;
