+++ packages/core/src/view/shape/node/SwimlaneShape.ts
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 import Shape from '../Shape';
-import Rectangle from '../Rectangle';
+import Rectangle from '../../geometry/Rectangle';
 import {
   DEFAULT_STARTSIZE,
   LINE_ARCSIZE,
