+++ packages/core/src/view/shape/node/TextShape.ts
@@ -30,10 +30,10 @@ import {
 } from '../../../util/Constants';
 import { getBoundingBox } from '../../../util/mathUtils';
 import { getAlignmentAsPoint } from '../../../util/styleUtils';
-import Point from '../Point';
+import Point from '../../geometry/Point';
 import AbstractCanvas2D from '../../canvas/AbstractCanvas2D';
 import Shape from '../Shape';
-import Rectangle from '../Rectangle';
+import Rectangle from '../../geometry/Rectangle';
 import CellState from '../../cell/CellState';
 import { htmlEntities, replaceTrailingNewlines, trim } from '../../../util/StringUtils';
 import { isNode } from '../../../util/domUtils';
