+++ packages/core/src/view/shape/node/TriangleShape.ts
@@ -16,65 +16,50 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import Shape from '../Shape';
-import Point from '../Point';
+import Point from '../../geometry/Point';
+import ActorShape from './ActorShape';
 import { LINE_ARCSIZE } from '../../../util/Constants';
-import Rectangle from '../Rectangle';
 import AbstractCanvas2D from '../../canvas/AbstractCanvas2D';
 
 /**
- * Extends {@link Shape} to implement a rhombus (aka diamond) shape.
+ * Implementation of the triangle shape.
  *
- * This shape is registered under `rhombus` in {@link CellRenderer} when using {@link Graph} or calling {@link registerDefaultShapes}.
+ * This shape is registered under `triangle` in {@link CellRenderer} when using {@link Graph} or calling {@link registerDefaultShapes}.
  *
  * @category Vertex Shapes
  */
-class RhombusShape extends Shape {
-  constructor(bounds: Rectangle, fill: string, stroke: string, strokewidth = 1) {
+class TriangleShape extends ActorShape {
+  constructor() {
     super();
-    this.bounds = bounds;
-    this.fill = fill;
-    this.stroke = stroke;
-    this.strokeWidth = strokewidth;
   }
 
   /**
    * Adds roundable support.
+   * @returns {boolean}
    */
-  // isRoundable(): boolean;
-  isRoundable(): boolean {
+  isRoundable() {
     return true;
   }
 
   /**
-   * Generic painting implementation.
+   * Draws the path for this shape.
    * @param {mxAbstractCanvas2D} c
    * @param {number} x
    * @param {number} y
    * @param {number} w
    * @param {number} h
    */
-  paintVertexShape(c: AbstractCanvas2D, x: number, y: number, w: number, h: number) {
-    const hw = w / 2;
-    const hh = h / 2;
-
+  redrawPath(c: AbstractCanvas2D, x: number, y: number, w: number, h: number) {
     const arcSize = (this.style?.arcSize ?? LINE_ARCSIZE) / 2;
 
-    c.begin();
     this.addPoints(
       c,
-      [
-        new Point(x + hw, y),
-        new Point(x + w, y + hh),
-        new Point(x + hw, y + h),
-        new Point(x, y + hh),
-      ],
+      [new Point(0, 0), new Point(w, 0.5 * h), new Point(0, h)],
       this.isRounded,
       arcSize,
       true
     );
-    c.fillAndStroke();
   }
 }
 
-export default RhombusShape;
+export default TriangleShape;
