+++ packages/core/src/view/shape/stencil/StencilShapeRegistry.ts
