+++ packages/core/src/view/shape/stencil/register.ts
