+++ packages/core/src/view/style/marker/EdgeMarkerRegistry.ts
@@ -24,7 +24,7 @@ import type {
 } from '../../../types';
 import type AbstractCanvas2D from '../../canvas/AbstractCanvas2D';
 import type Point from '../../geometry/Point';
-import type Shape from '../../geometry/Shape';
+import type Shape from '../../shape/Shape';
 import { BaseRegistry } from '../../../internal/BaseRegistry';
 
 class EdgeMarkerRegistryImpl
